Update to react router 2.0, Route whole DemoPage
