--- DemoPage/examples/Wizard/AppHandler.js
@@ -1,15 +1,0 @@
-'use strict'
-
-import React  from 'react'
-import Router from 'react-router'
-var RouteHandler = Router.RouteHandler;
-
-export default React.createClass({
-  displayName: 'AppHandler',
-
-  mixins: [Router.State],
-
-  render: function() {
-    return <RouteHandler/>;
-  }
-});
