+++ server.js
@@ -4,10 +4,12 @@ import express from 'express'
 import compression from 'compression'
 import morgan from 'morgan'
 import React from 'react'
+import { renderToString } from 'react-dom/server'
+import { match, RouterContext } from 'react-router'
+import DemoPageRoutes from './DemoPage/Routes'
 
 const app = express()
 
-const ReactApp = React.createFactory(require('./DemoPage'))
 const staticPath = __dirname + '/dist/'
 const port = Number(process.env.PORT || 8000)
 
@@ -15,9 +17,11 @@ app.use(compression())
 app.use(morgan())
 
 const sendStaticFile = name =>
-  (_, res) => {
-    let content = React.renderToString(ReactApp({}))
-    res.render(staticPath + name, { content })
+  (req, res) => {
+    match({ routes: DemoPageRoutes, location: req.path }, (error, redirectLocation, renderProps) => {
+      let content = renderToString(React.createElement(RouterContext, renderProps))
+      res.render(staticPath + name, { content })
+    })
   }
 
 app.use(express.static(staticPath))
